회원가입, 로그인 로직을 auth 파일로 분리
