+++ src/main/java/com/aloc/aloc/auth/controller/AuthController.java
@@ -1,0 +1,32 @@
+package com.aloc.aloc.auth.controller;
+
+import com.aloc.aloc.auth.dto.LoginRequestDto;
+import com.aloc.aloc.auth.dto.LoginResponseDto;
+import com.aloc.aloc.auth.dto.RegisterRequestDto;
+import com.aloc.aloc.auth.service.AuthService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/auth")
+public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/register")
+    public ResponseEntity<String> register(@RequestBody RegisterRequestDto registerRequestDto) {
+        authService.register(registerRequestDto);
+        return ResponseEntity.ok("회원가입 성공");
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponseDto> login(@RequestBody LoginRequestDto loginRequestDto) {
+        LoginResponseDto loginResponseDto = authService.login(loginRequestDto);
+        return ResponseEntity.ok(loginResponseDto);
+    }
+}
