+++ src/main/java/com/aloc/aloc/user/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.aloc.aloc.user.controller;
 
-import com.aloc.aloc.user.dto.LoginRequestDto;
-import com.aloc.aloc.user.dto.LoginResponseDto;
-import com.aloc.aloc.user.dto.RegisterRequestDto;
+import com.aloc.aloc.auth.dto.LoginRequestDto;
+import com.aloc.aloc.auth.dto.LoginResponseDto;
+import com.aloc.aloc.auth.dto.RegisterRequestDto;
 import com.aloc.aloc.user.dto.UserResponseDto;
 import com.aloc.aloc.user.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -18,20 +18,10 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping("/register")
-    public ResponseEntity<String> register(@RequestBody RegisterRequestDto registerRequestDto) {
-        userService.register(registerRequestDto);
-        return ResponseEntity.ok("회원가입 성공");
-    }
-
     @GetMapping
     public List<UserResponseDto> getAllUsers() {
         return userService.getAllUsers();
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponseDto> login(@RequestBody LoginRequestDto loginRequestDto) {
-        LoginResponseDto loginResponseDto = userService.login(loginRequestDto);
-        return ResponseEntity.ok(loginResponseDto);
-    }
+
 }
